refactor: move date&time converter to each CouponMapper.kt, CouponUiModelConverter.kt
- remove util package
