+++ app/src/main/java/woowacourse/shopping/data/mapper/CouponMapper.kt
@@ -11,8 +11,12 @@ import woowacourse.shopping.domain.model.coupon.Coupon
 import woowacourse.shopping.domain.model.coupon.FixedDiscountCoupon
 import woowacourse.shopping.domain.model.coupon.FreeShippingCoupon
 import woowacourse.shopping.domain.model.coupon.PercentageDiscountCoupon
-import woowacourse.shopping.util.convertStringToLocalDate
-import woowacourse.shopping.util.convertStringToLocalTime
+import java.time.LocalDate
+import java.time.LocalTime
+import java.time.format.DateTimeFormatter
+
+private const val DATE_PARSE_PATTERN = "yyyy-MM-dd"
+private const val TIME_PARSE_PATTERN = "HH:mm:ss"
 
 fun CouponDto.toDomainModel(): Coupon {
     return when (this) {
@@ -66,4 +70,12 @@ fun CouponDto.toDomainModel(): Coupon {
             )
         }
     }
+}
+
+fun convertStringToLocalDate(date: String): LocalDate {
+    return LocalDate.parse(date, DateTimeFormatter.ofPattern(DATE_PARSE_PATTERN))
+}
+
+fun convertStringToLocalTime(time: String): LocalTime {
+    return LocalTime.parse(time, DateTimeFormatter.ofPattern(TIME_PARSE_PATTERN))
 }
